// interfaces/common.ts
export interface FitnessPlanData {
  fitnessPlan: string;
  lowIntensityNeeded: number;
  lowIntensityCompleted: number;
  lowIntensityPending: number;
  mediumIntensityNeeded: number;
  mediumIntensityCompleted: number;
  mediumIntensityPending: number;
  highIntensityNeeded: number;
  highIntensityCompleted: number;
  highIntensityPending: number;
}